Implement RtlRandom (which despite its name is located in Xam, not xboxkrnl, which is why its not in xboxkrnl_rtl) Stub XamVoiceSubmitPacket Stubs for __CAP_Start_Profiling/End_Profiling/Enter_Function/Exit_Function add a note about io_status_block in NtDeviceIoControlFile, change the param's type move the X_IOCTL_ constants from NtDeviceIoControlFile's body into xbox.h add X_STATUS_INVALID_IMAGE_FORMAT to xbox.h Implement XexLoadImageHeaders Much more correct version of IoCreateDevice, properly initializes/allocates device extension data Stub version of IoDeleteDevice Open the quickstart guide in the browser the first time a user opens the emulator Add some persistent flags that are stored in the registry on windows, not in the config. These are just used to indicate whether one-time tasks have run, like showing the quickstart guide. On non-windows platforms a default value is returned that skips any of these tasks so they don't run every single time.
If the user opens a .iso file, show a warning telling them that we do not condone or support piracy. If the user closes the messagebox within two seconds set a sticky flag that will show them the warning every time instead. Otherwise the warning is never shown again. The Beep function is used to spook them a bit, to hopefully make them less likely to skip the message
If a user opens an archive file, which we do not support, explain to them that they're dumb.

Removed messages intended to "catch" pirates.  If they're out there, we don't want to know about them. A huge chunk of the discourse on our discord is now about piracy. Hopefully these changes will deter them from coming to us for now, but some of these messages may backfire