[App] Implement per-game configs.
User visible interface is the same, i.e. config/GAMEID.config.toml
but implementation uses the fact that Edge spawns a separate
game process to parse the overrides in the UI process and pass them
to the game process as commandline args.

Removing all the previous infrastructure with callbacks, etc. that
was attempting to work around the basic limitation of needing to
know the overrides before doing any other init.